Widen 16-bit peak accumulators to int
The 16-bit peak code stored abs(*sample_pointer) into an int16_t. For
samples at -32768, abs() returns 32768, which overflows int16_t and
typically wraps back to -32768 — silently dropping the loudest peaks.

Discovered by cross-checking against an independent pure-Ruby port:
on real audio that clips at the int16_t minimum, ~2% of output points
were 1 unit lower than they should have been. Compare with the Ruby
implementation now produces byte-identical output on
NR_Success.16.wav with --peak --scale 255.

The 8-bit peak path uses int16_t but is safe (abs(int8_t) maxes at
128). The 24-bit path already uses int32_t. Only 16-bit peak was
affected.

+++ waveform.c
@@ -473,7 +473,8 @@ int waveform_1_channel_16_bit_peak(int16_t *samples, int sample_group_size, int 
 
 	int16_t *sample_pointer = samples;
 
-	int16_t sample_point, peak_0=0;
+	/* int, not int16_t: abs(-32768) is 32768, which doesn't fit in int16_t. */
+	int sample_point, peak_0=0;
 
 	if (same_endianness) {
 		for (j=0 ; j<sample_group_size ; j++) {
@@ -506,7 +507,8 @@ int waveform_2_channel_16_bit_peak(int16_t *samples, int sample_group_size, int 
 
 	int16_t *sample_pointer = samples;
 
-	int16_t sample_point, peak_0=0, peak_1=0;
+	/* int, not int16_t: abs(-32768) is 32768, which doesn't fit in int16_t. */
+	int sample_point, peak_0=0, peak_1=0;
 
 	if (same_endianness) {
 		for (j=0 ; j<sample_group_size ; j++) {
